Make more of the ToDotNet unit tests file-based.
