+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/ThrowsIfReferenceDoesNotSpecifyADefinition/Schema.json
@@ -1,0 +1,13 @@
+
+{
+  "type": "object",
+  "properties": {
+    "p": {
+      "$ref": "#/notDefinitions/p"
+    }
+  },
+  "notDefinitions": {
+    "p": {
+    }
+  }
+}
