+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/ThrowsIfReferencedDefinitionDoesNotExist/Schema.json
@@ -1,0 +1,13 @@
+
+{
+  "type": "object",
+  "properties": {
+    "p": {
+      "$ref": "#/definitions/nonExistentDefinition"
+    }
+  },
+  "definitions": {
+    "p": {
+    }
+  }
+}
